+++ cinetica/__init__.py
@@ -2,7 +2,7 @@
 Cinetica - Una librería para cálculos de cinemática
 """
 
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 from .units import ureg, Q_